+++ src/legoapi/render/core/render.cpp
@@ -1285,9 +1285,9 @@ void DrawGameState(float x, float y, i32 highlight, i32 slot) {
     u8 blue = MENUNORMALB;
     if (highlight != 0 && TestForController() != 0) {
         if (menu_pulsate > 0.0f) {
-            red = static_cast<u8>(static_cast<i32>(MENUFLASH0R * menu_pulsate + MENUFLASH1R * (1.0f - menu_pulsate)));
-            green = static_cast<u8>(static_cast<i32>(MENUFLASH0G * menu_pulsate + MENUFLASH1G * (1.0f - menu_pulsate)));
-            blue = static_cast<u8>(static_cast<i32>(MENUFLASH0B * menu_pulsate + MENUFLASH1B * (1.0f - menu_pulsate)));
+            red = static_cast<u8>(MENUFLASH0R * menu_pulsate + MENUFLASH1R * (1.0f - menu_pulsate));
+            green = static_cast<u8>(MENUFLASH0G * menu_pulsate + MENUFLASH1G * (1.0f - menu_pulsate));
+            blue = static_cast<u8>(MENUFLASH0B * menu_pulsate + MENUFLASH1B * (1.0f - menu_pulsate));
         } else if (menu_flash != 0) {
             red = MENUFLASH0R;
             green = MENUFLASH0G;
@@ -1298,9 +1298,9 @@ void DrawGameState(float x, float y, i32 highlight, i32 slot) {
             blue = MENUFLASH1B;
         }
     } else if (menu_pulse > 0.0f) {
-        red = static_cast<u8>(static_cast<i32>(MENUFLASH0R * menu_pulse + MENUNORMALR * (1.0f - menu_pulse)));
-        green = static_cast<u8>(static_cast<i32>(MENUFLASH0G * menu_pulse + MENUNORMALG * (1.0f - menu_pulse)));
-        blue = static_cast<u8>(static_cast<i32>(MENUFLASH0B * menu_pulse + MENUNORMALB * (1.0f - menu_pulse)));
+        red = static_cast<u8>(MENUFLASH0R * menu_pulse + MENUNORMALR * (1.0f - menu_pulse));
+        green = static_cast<u8>(MENUFLASH0G * menu_pulse + MENUNORMALG * (1.0f - menu_pulse));
+        blue = static_cast<u8>(MENUFLASH0B * menu_pulse + MENUNORMALB * (1.0f - menu_pulse));
     }
     SmartTextEx(game_name, x, y, 1.0f, MENUTEXTSCALE, MENUTEXTSCALE, MENUTEXTSCALE, 4, red, green, blue, 0.45f, 1, NULL,
                 0, MenuA);
@@ -1821,9 +1821,8 @@ void DrawMiniKitCount(float position, float scale, i32 count, i32 maximum) {
     const f32 y = (KITPOSY - KITPOS2Y) * blend + KITPOS2Y;
     WORLDINFO_s *world = WorldInfo_CurrentlyActive();
     if (world->lev_objs[model].active != 0) {
-        const u16 rotation =
-            static_cast<u16>(static_cast<i32>(NuFmod(GameTimer.time_elapsed, 4.0f) * 0.25f * 65536.0f));
-        const u16 tilt = static_cast<u16>(static_cast<i32>(1820.0f * NuTrigTable[rotation & 0x7fff]));
+        const u16 rotation = static_cast<u16>(NuFmod(GameTimer.time_elapsed, 4.0f) * 0.25f * 65536.0f);
+        const u16 tilt = static_cast<u16>(1820.0f * NuTrigTable[rotation & 0x7fff]);
         const f32 size = scale * PANEL_MINIKITSCALE;
         DrawPanel3DObjectNoAlpha(x, PANEL_MINIKITY + y, 1.0f, size, size, size, tilt, rotation, 0,
                                  &world->lev_objs[model].special, 2);
@@ -3310,7 +3309,7 @@ static void DrawParaphernalia(GameObject_s *object) {
                 NUVEC scaling = {scale, scale, scale};
                 NUMTX matrix;
                 NuMtxSetScale(&matrix, &scaling);
-                NuMtxRotateX(&matrix, static_cast<u16>(static_cast<i32>(-65536.0f * fraction)));
+                NuMtxRotateX(&matrix, static_cast<u16>(-65536.0f * fraction));
                 NuMtxRotateY(&matrix, object->apiobj.field_0x276);
                 NuMtxTranslate(&matrix, reinterpret_cast<NUVEC *>(&joints[0].m30));
                 matrix.m31 += 0.2f * NuTrigTable[(static_cast<i32>(fraction * 32768.0f) >> 1) & 0x7fff];
